[Auto Generated] 0.1.26

+++ Sources/code/application/models/ShipmentTotalDetailsAppModelClass.swift
@@ -7,11 +7,11 @@ public extension ApplicationClient {
          Used By: Order
      */
     class ShipmentTotalDetails: Codable {
-        public var sizes: Double?
+        public var sizes: Int?
 
         public var totalPrice: Double?
 
-        public var pieces: Double?
+        public var pieces: Int?
 
         public enum CodingKeys: String, CodingKey {
             case sizes
@@ -21,7 +21,7 @@ public extension ApplicationClient {
             case pieces
         }
 
-        public init(pieces: Double? = nil, sizes: Double? = nil, totalPrice: Double? = nil) {
+        public init(pieces: Int? = nil, sizes: Int? = nil, totalPrice: Double? = nil) {
             self.sizes = sizes
 
             self.totalPrice = totalPrice
@@ -33,7 +33,7 @@ public extension ApplicationClient {
             let container = try decoder.container(keyedBy: CodingKeys.self)
 
             do {
-                sizes = try container.decode(Double.self, forKey: .sizes)
+                sizes = try container.decode(Int.self, forKey: .sizes)
 
             } catch DecodingError.typeMismatch(let type, let context) {
                 print("Type '\(type)' mismatch:", context.debugDescription)
@@ -49,7 +49,7 @@ public extension ApplicationClient {
             } catch {}
 
             do {
-                pieces = try container.decode(Double.self, forKey: .pieces)
+                pieces = try container.decode(Int.self, forKey: .pieces)
 
             } catch DecodingError.typeMismatch(let type, let context) {
                 print("Type '\(type)' mismatch:", context.debugDescription)
